+++ circleguard/comparer.py
@@ -6,7 +6,7 @@ import math
 import numpy as np
 
 from circleguard.loadable import Replay
-from circleguard.enums import Mod
+from circleguard.mod import Mod
 from circleguard.exceptions import InvalidArgumentsException, CircleguardException
 import circleguard.utils as utils
 from circleguard.result import StealResult
